New version 1.05 for the reference concepts structure. New version 1.02 for the geo module.
These new versions will lead to the overall release of the new UMBEL version 1.05 (which will be tagged as such).

All previous version: 1.01, 1.02, 1.03 and 1.04 should be deprecated and not used anymore.

These new versions fixes multiple issues we were perpetuating version after version. These two new files are based on UMBEL v1.00, and which as been fixed a bit (a few hundred of RC classes where not punned as RefConcepts).